Docs: move get data api from pr to home page
