+++ client/src/componnents/ProtectedRoute.jsx
@@ -1,45 +1,5 @@
-import { useSelector, useDispatch } from "react-redux";
-import { useEffect } from "react";
-import { Navigate, useNavigate } from "react-router-dom";
-import axios from "axios";
-import { setUser } from "../redux/userSlice";
-import { showLoading, hideLoading } from "../redux/alertSlice";
+import { Navigate } from "react-router-dom";
 function ProtectedRoute(props) {
-  const { user } = useSelector((state) => state.user);
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-  const getUser = async () => {
-    // token validation check  before login
-    try {
-      dispatch(showLoading());
-      const response = await axios.post(
-        "/api/user/get-user-info-by-id",
-        { token: localStorage.getItem("token") },
-        {
-          headers: {
-            Authorization: `Bearer ${localStorage.getItem("token")}`,
-          },
-        }
-      );
-      dispatch(hideLoading());
-      if (response.data.success) {
-        dispatch(setUser(response.data.data)); // passing the user data redux
-      } else {
-        localStorage.clear();
-        navigate("/login");
-      }
-    } catch (error) {
-      dispatch(hideLoading());
-      localStorage.clear(); // this when wrong token pass
-      navigate("/login");
-    }
-  };
-
-  useEffect(() => {
-    if (!user) {
-      getUser();
-    }
-  }, [user]);
   //starting web page from login
   if (localStorage.getItem("token")) {
     return props.children;
